Assume first metadata value is GEN8 version
We will need to work with different versions of the format, and this value
seems to be consistent among different games made with the same version of
Game Maker Studio, but it seems to differ if the game was made with another
version of Game Maker Studio.

So let's assume this is some kind of GEN8 version ... thing.